Change bcash with Bitcoin Cash
Bcash is a despective name made to create confusion and deviate people to the subreddit r/bcash where people is trolled. It is wrote in hundreds of tweet to known Bitcoin Cash personalities trying to bother them. "Bitcoin" is not a registered mark and it could freely used in a name but lot of BTC supports don't accept it. Probably nobody that likes Bitcoin Cash would donate to a place wich call the currency "bcash".